added timeline breakpoint
This is for the home and hospital tests. I want to keep everything in
the same place when I’m running tests on development mockups. This
media query conditional will remove arrows pointing to the right and
make them point down when the viewport is set to 685 pixels or less.
The opposite happens when it’s 685 pixels or more.